+++ README.md
@@ -50,7 +50,7 @@
 - [ ] 仮想経済データの動的変化処理
 
 ## 使い方
-ブラウザで `public/index.html` を開き、`開始` ボタンをクリックするとゲームがスタートします。
+ブラウザで `public/start_screen.html` を開き、`開始` ボタンをクリックするとゲームがスタートします。
 
 ## 🧪 テスト実行方法
