--- public/game.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html lang="ja">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>ゲーム画面</title>
-    <link rel="stylesheet" href="styles.css">
-</head>
-<body>
-    <div id="game-container">
-        <h1>ここからゲームが始まります</h1>
-        <p>コンテンツは今後追加予定です。</p>
-    </div>
-</body>
-</html>
