+++ public/game_screen.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>ゲーム画面</title>
-    <link rel="stylesheet" href="styles.css">
+    <link rel="stylesheet" href="game_screen.css">
 </head>
 <body>
     <div id="game-container">
