+++ public/start_screen.css
