+++ public/start_screen.html
@@ -8,9 +8,9 @@
   <!-- Tailwind CDN -->
   <script src="https://cdn.tailwindcss.com"></script>
   <!-- 外部CSSを読み込み -->
-  <link rel="stylesheet" href="start.css" />
+  <link rel="stylesheet" href="start_screen.css" />
   <!-- 外部JavaScriptを読み込み -->
-  <script defer src="start.js"></script>
+  <script defer src="start_screen.js"></script>
 </head>
   <body class="h-screen w-screen flex items-center justify-center select-none">
     <!-- 中央にゲームタイトルのみ表示し、画面全体をタップで遷移 -->
